added an env variable THEANO_NOCLEANUP. When true, we don't delete any generated file.

+++ theano/gof/cmodule.py
@@ -386,7 +386,7 @@ class ModuleCache(object):
             try:
                 module = fn(location=location)  # WILL FAIL FOR BAD C CODE
             except Exception, e:
-                shutil.rmtree(location)
+                _rmtree(location)
                 #try:
                 #except Exception, ee:
                     #error('failed to cleanup location', location, ee)
@@ -515,7 +515,11 @@ class ModuleCache(object):
 
 def _rmtree(parent):
     try:
-        shutil.rmtree(parent)
+        if not os.getenv('THEANO_NOCLEANUP',0):
+            print "cmodule.py do cleanup"
+            shutil.rmtree(parent)
+        else:
+            print "cmodule.py don't cleanup"
     except Exception, e:
         try:
             # mark this directory for deletion by a future refresh()
